Moved and renamed the new test case.
