+++ tests/forms_tests/field_tests/test_base.py
@@ -1,4 +1,4 @@
-from django.forms import Field
+from django.forms import ChoiceField, Field, Form, Select
 from django.test import SimpleTestCase
 
 
@@ -20,3 +20,17 @@ class BasicFieldsTests(SimpleTestCase):
 
         f = ComplexField()
         self.assertTrue(f.class_a_var)
+
+    def test_field_deepcopies_widget_instance(self):
+        class CustomChoiceField(ChoiceField):
+            widget = Select(attrs={'class': 'my-custom-class'})
+
+        class TestForm(Form):
+            field1 = CustomChoiceField(choices=[])
+            field2 = CustomChoiceField(choices=[])
+
+        f = TestForm()
+        f.fields['field1'].choices = [('1', '1')]
+        f.fields['field2'].choices = [('2', '2')]
+        self.assertEqual(f.fields['field1'].widget.choices, [('1', '1')])
+        self.assertEqual(f.fields['field2'].widget.choices, [('2', '2')])
